+++ examples/example0r.cc
@@ -14,28 +14,28 @@ int main()
 
   fftw::maxthreads=get_max_threads();
 
-  unsigned int n=4;
-  unsigned int np=n/2+1;
+  size_t n=4;
+  size_t np=n/2+1;
   double *f=FFTWdouble(n);
   Complex *g=FFTWComplex(np);
 
   rcfft1d Forward(n,f,g);
   crfft1d Backward(n,g,f);
 
-  for(unsigned int i=0; i < n; i++) f[i]=i;
+  for(size_t i=0; i < n; i++) f[i]=i;
 
   cout << "\ninput (" << n << " doubles):" << endl;
-  for(unsigned int i=0; i < n; i++) cout << f[i] << (i!=n-1 ? " " : "\n");
+  for(size_t i=0; i < n; i++) cout << f[i] << (i!=n-1 ? " " : "\n");
 
   Forward.fft(f,g);
 
   cout << "\noutput (" << np << " complex):" << endl;
-  for(unsigned int i=0; i < np; i++) cout << g[i] <<  (i!=np-1 ? " " : "\n");
+  for(size_t i=0; i < np; i++) cout << g[i] <<  (i!=np-1 ? " " : "\n");
 
   Backward.fftNormalized(g,f);
 
   cout << "\ntransformed back:" << endl;
-  for(unsigned int i=0; i < n; i++) cout << f[i] <<  (i!=n-1 ? " " : "\n");
+  for(size_t i=0; i < n; i++) cout << f[i] <<  (i!=n-1 ? " " : "\n");
 
   deleteAlign(g);
   deleteAlign(f);
