+++ examples/example2.cc
@@ -14,7 +14,7 @@ int main()
   cout << "2D complex to complex in-place FFT" << endl;
   fftw::maxthreads=get_max_threads();
 
-  unsigned int nx=4, ny=4;
+  size_t nx=4, ny=4;
   size_t align=sizeof(Complex);
 
   array2<Complex> f(nx,ny,align);
@@ -22,8 +22,8 @@ int main()
   fft2d Forward(-1,f);
   fft2d Backward(1,f);
 
-  for(unsigned int i=0; i < nx; i++)
-    for(unsigned int j=0; j < ny; j++)
+  for(size_t i=0; i < nx; i++)
+    for(size_t j=0; j < ny; j++)
       f(i,j)=Complex(i,j);
 
   cout << "\ninput:\n" << f;
